// src/app/types/payments/transactions/payment-transaction.types.ts
export enum PaymentMethodKind {
    BankTransfer = "bank_transfer",
    Cash = "cash",
    Cheque = "cheque",
    Card = "card",
    Gateway = "gateway",
}
// ...
export interface PaymentTransactionCreateInputDto {
    bankAccountAlias: string;// Bank account

    amount: number;
    currencyCode: string;

    method: PaymentMethodKind;

    externalRef?: string;
    transactionAt: ISODateString;

    notes?: string;
}